Extension project file cleanup
Also done some prep work for future extension work if I am unable to find a way to get the grammar to work on .cs files. In that case I will try setting vs up to change the extension to .csasm whenever the cursor enters an asm string region and back afterwards.