Split asm structures into a different project
Setup easier debugging of the source generator based on https://github.com/JoanComasFdz/dotnet-how-to-debug-source-generator-vs2022 .
Initially I was having issues that appeared to be related to using the source generator project as a regular library as well and so that part has be split out, this may not have actually been necessary.

Removed the option to use either the library(and everything else) on anything other than x86_64 since it won't run anyway. This may be extended to include x86 at some point. Additionally this limitation may be removed for components that are mostly related to development to allow building applications using the library from other arches.