Fix crashes when using variables in string format
This did reveal that the use of variables will cause the two formats to give different guids and so the source generator will emit two separate function, even though they are the same.

The crash was from the source generator using the value of a variable while at runtime the guid function was using the variable name itself and so was failing to find the matching function.