Add partial support for custom label names/multiple labels
This method has some shortcomings as it requires the name of the variable representing the label to have the same name as the label itself for both the hashes to agree. Additionally, if a label name begins with __ then the hashes won't match either since that is the prefix that Iced uses to indicate that a label has not been renamed.